Code refactoring for better modules responsibilities
