+++ src/request/soap.js
@@ -1,7 +1,14 @@
 import md5 from 'blueimp-md5';
+import assign from 'object-assign';
 
 import {get as getToken} from '../token';
 import * as request from '../request/native';
+
+import * as settings from '../settings';
+import {getDefault} from '../quality';
+
+const {TRANSLATION} = settings.params;
+const {ANY, RUSSIAN, SUBTITLES} = settings.values[TRANSLATION];
 
 export function get(url) {
 	return request.get(url, headers());
@@ -68,6 +75,40 @@ export function getTVShowSeasons(sid) {
 
 export function getTVShowSeason(sid, id) {
 	return getTVShowSeasons(sid).then(seasons => seasons.filter(season => season.id === id)[0]);
+}
+
+export function getResolvedSeasonEpisodes(season) {
+	let translationSettings = settings.get(TRANSLATION);
+	let {
+		episodes: episodesList = [],
+		subtitles: subtitlesList = [],
+	} = season;
+
+	let episodes = episodesList
+		.filter(Boolean)
+		.map(getDefault);
+
+	let subtitles = subtitlesList
+		.filter(Boolean)
+		.map(getDefault)
+		.map((episode, i) => assign({}, episode, {
+			hasSubtitles: true,
+		}));
+
+	let episodesIds = episodes.map(({episode}) => episode);
+
+	if (translationSettings === RUSSIAN || !subtitles.length) {
+		return episodes;
+	} else if (translationSettings === SUBTITLES) {
+		return subtitles;
+	} else if (translationSettings === ANY) {
+		return subtitles.map(episode => {
+			let {episode: id} = episode;
+			let episodesIndex = episodesIds.indexOf(id);
+			return ~episodesIndex ? episodes[episodesIndex] : episode;
+		});
+	}
+	return [];
 }
 
 export function markEpisodeAsWatched(eid) {
